+++ lua/ergoterm/commandline.lua
@@ -23,6 +23,7 @@ local p = {
 ---@field trim boolean?
 ---@field new_line boolean?
 ---@field auto_scroll boolean?
+---@field bang_target boolean?
 ---@field watch_files boolean?
 ---@field persist_mode boolean?
 ---@field selectable boolean?
@@ -63,6 +64,7 @@ function M.parse(args)
       trim = true,
       new_line = true,
       auto_scroll = true,
+      bang_target = true,
       watch_files = true,
       persist_mode = true,
       selectable = true,
@@ -217,6 +219,8 @@ M._all_options = {
 
   auto_scroll = M._boolean_options,
 
+  bang_target = M._boolean_options,
+
   watch_files = M._boolean_options,
 
   persist_mode = M._boolean_options,
@@ -238,6 +242,7 @@ M._term_new_options = {
   layout = M._all_options.layout,
   name = M._all_options.name,
   auto_scroll = M._all_options.auto_scroll,
+  bang_target = M._all_options.bang_target,
   watch_files = M._all_options.watch_files,
   persist_mode = M._all_options.persist_mode,
   selectable = M._all_options.selectable,
@@ -251,6 +256,7 @@ M._term_update_options = {
   layout = M._all_options.layout,
   name = M._all_options.name,
   auto_scroll = M._all_options.auto_scroll,
+  bang_target = M._all_options.bang_target,
   watch_files = M._all_options.watch_files,
   persist_mode = M._all_options.persist_mode,
   selectable = M._all_options.selectable,